Factor out all the redundancy in the Optic type
Now the Optic type captures exactly all the variations that I know
about at this moment, with none of the common stuff repeated. That
means everything has to be calculated from these values.

The intention is that these values will be easy to compose with each
other by directly comparing the component parts, with any need to
analyze Haskell types.